Introduce SignedFile super class
