+++ signify/exceptions.py
@@ -1,6 +1,11 @@
 class SignifyError(Exception):
     """Base class for all errors generated by Signify"""
 
+    def __eq__(self, other: object) -> bool:
+        return type(self) == type(other) and getattr(self, "args") == getattr(
+            other, "args"
+        )
+
 
 class ParseError(SignifyError):
     pass
